refactor: extract page logic into custom hooks and auth helpers
