+++ app/console/tokens/page.tsx
@@ -1,6 +1,6 @@
 /**
  * @author Claude
- * @version 1.2
+ * @version 1.3
  * @date 2026/5/22
  *
  * Tokens page with SSR auth check and token list preloading.
@@ -8,10 +8,8 @@
 import TokensPage from '@/app/console/tokens/page.client';
 import { getTranslations } from 'next-intl/server';
 import type { Metadata } from 'next';
-import { headers } from 'next/headers';
-import { redirect } from 'next/navigation';
 
-import { fetchMe } from '@/services/sso/server';
+import { requireAuth } from '@/app/lib/auth';
 import { fetchTokenList } from '@/services/token/server';
 
 export const generateMetadata = async (): Promise<Metadata> => {
@@ -22,15 +20,7 @@ export const generateMetadata = async (): Promise<Metadata> => {
 };
 
 const Page = async () => {
-	const me = await fetchMe();
-	if (!me) {
-		const headersList = await headers();
-		const host = headersList.get('host') || 'localhost:3000';
-		const protocol = headersList.get('x-forwarded-proto') || 'https';
-		const from = `${protocol}://${host}/console/tokens`;
-		redirect(`/login?from=${encodeURIComponent(from)}`);
-	}
-
+	await requireAuth('/console/tokens');
 	const initialTokens = await fetchTokenList();
 	return <TokensPage initialTokens={initialTokens} />;
 };
