+++ app/login/store.ts
@@ -1,6 +1,6 @@
 /**
  * @author Claude
- * @version 1.0
+ * @version 1.1
  * @date 2026/5/22
  *
  * Jotai atoms for the /login page.
@@ -8,6 +8,7 @@
  * - `fromAtom`        — URL to redirect to after login (from searchParams).
  * - `loginMeAtom`     — set by QRLoginView / PasskeyLoginView on success.
  * - `mobileAtom`      — whether the current device is mobile.
+ * - `stateAtom`       — OAuth2 state for CSRF protection, generated on mount.
  * - `deepLinkUrlAtom` — derived `ham://` deep-link URL for mobile app login.
  */
 
@@ -15,6 +16,16 @@ import { atom } from 'jotai';
 
 import { MeResponse } from '@/services/sso/api';
 import { buildSsoAuthorizeDeepLink } from '@/services/sso/deepLink';
+
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
+/**
+ * Fixed redirect_uri for app OAuth2 callback.
+ * Must be registered with the backend's allowed redirect URIs.
+ */
+export const APP_CALLBACK_PATH = '/login/callback';
 
 // ---------------------------------------------------------------------------
 // Atoms
@@ -29,14 +40,18 @@ export const loginMeAtom = atom<MeResponse | null>(null);
 /** Whether the current device is mobile. Set on mount. */
 export const mobileAtom = atom(false);
 
-/** Derived: builds the `ham://` deep-link URL from current `fromAtom`. */
+/** OAuth2 state parameter for CSRF protection. Generated on mount. */
+export const stateAtom = atom<string>('');
+
+/** Derived: builds the `ham://` deep-link URL with fixed redirectUri and state. */
 export const deepLinkUrlAtom = atom<string>((get) => {
-	const from = get(fromAtom);
-	if (!from) return '';
+	const state = get(stateAtom);
+	if (!state) return '';
+	const origin = typeof window !== 'undefined' ? window.location.origin : '';
 	return buildSsoAuthorizeDeepLink({
 		appId: process.env.NEXT_PUBLIC_CONSOLE_CLIENT_ID ?? '',
 		scope: [],
-		state: '',
-		redirectUri: from,
+		state,
+		redirectUri: `${origin}${APP_CALLBACK_PATH}`,
 	});
 });
