+++ app/sso-authorize/useConsent.ts
@@ -1,0 +1,138 @@
+/**
+ * @author Claude
+ * @version 1.0
+ * @date 2026/5/22
+ *
+ * Custom hook for SSO consent view logic.
+ * Handles consent info fetching, scope selection, confirm/reject/switch account.
+ */
+
+'use client';
+
+import { useAtomValue } from 'jotai';
+import { useTranslations } from 'next-intl';
+import { useCallback, useEffect, useState } from 'react';
+import toast from 'react-hot-toast';
+
+import { ApiError, ConsentInfoResponse, WebAuthApi } from '@/services/sso/api';
+import { paramsAtom, stageAtom } from '@/app/sso-authorize/store';
+
+export interface UseConsentReturn {
+	me: { user_id: string; nickname?: string; avatar_url?: string } | null;
+	info: ConsentInfoResponse | null;
+	error: string | null;
+	submitting: boolean;
+	checkedScopes: string[];
+	setCheckedScopes: (scopes: string[]) => void;
+	onSwitchAccount: () => Promise<void>;
+	bail: (oauthError: 'access_denied' | 'server_error') => void;
+	confirm: (nonce: string) => Promise<void>;
+}
+
+export const useConsent = (): UseConsentReturn => {
+	const params = useAtomValue(paramsAtom)!;
+	const stage = useAtomValue(stageAtom);
+	const me = stage.kind === 'consent' ? stage.me : null;
+	const t = useTranslations('sso.consent');
+	const [info, setInfo] = useState<ConsentInfoResponse | null>(null);
+	const [error, setError] = useState<string | null>(null);
+	const [submitting, setSubmitting] = useState(false);
+	const [checkedScopes, setCheckedScopes] = useState<string[]>([]);
+
+	useEffect(() => {
+		if (!info) return;
+		setCheckedScopes(info.scopes.map((s) => s.scope));
+	}, [info]);
+
+	const onSwitchAccount = useCallback(async () => {
+		try {
+			await WebAuthApi.logout();
+		} finally {
+			const from = encodeURIComponent(
+				window.location.pathname + window.location.search
+			);
+			window.location.href = `/login?from=${from}`;
+		}
+	}, []);
+
+	const bail = useCallback(
+		(oauthError: 'access_denied' | 'server_error') => {
+			try {
+				const url = new URL(params.redirectUri);
+				url.searchParams.set('error', oauthError);
+				if (params.state) {
+					url.searchParams.set('state', params.state);
+				}
+				window.location.replace(url.toString());
+			} catch {
+				setError(
+					oauthError === 'server_error' ? t('serverError') : t('accessDenied')
+				);
+			}
+		},
+		[params.redirectUri, params.state, t]
+	);
+
+	const confirm = useCallback(
+		async (nonce: string) => {
+			setSubmitting(true);
+			try {
+				const resp = await WebAuthApi.consentConfirm({
+					client_id: params.appId,
+					scope: checkedScopes,
+					redirect_uri: params.redirectUri,
+					state: params.state,
+					nonce,
+				});
+				window.location.replace(resp.redirect_url);
+			} catch (e) {
+				if (e instanceof ApiError) {
+					toast.error(e.message || t('submitFailed'));
+				} else {
+					toast.error(t('submitFailed'));
+				}
+				bail('server_error');
+			} finally {
+				setSubmitting(false);
+			}
+		},
+		[bail, checkedScopes, params.appId, params.redirectUri, params.state, t]
+	);
+
+	useEffect(() => {
+		let cancelled = false;
+		WebAuthApi.consentInfo({
+			client_id: params.appId,
+			scope: params.scope,
+			redirect_uri: params.redirectUri,
+			state: params.state,
+		})
+			.then((resp) => {
+				if (cancelled) return;
+				setInfo(resp);
+			})
+			.catch((e: unknown) => {
+				if (cancelled) return;
+				if (e instanceof ApiError) {
+					setError(e.message || t('fetchFailed'));
+				} else {
+					setError(t('fetchFailed'));
+				}
+			});
+		return () => {
+			cancelled = true;
+		};
+	}, [params.appId, params.redirectUri, params.scope, params.state, t]);
+
+	return {
+		me,
+		info,
+		error,
+		submitting,
+		checkedScopes,
+		setCheckedScopes,
+		onSwitchAccount,
+		bail,
+		confirm,
+	};
+};
