+++ eslint.config.ts
@@ -13,6 +13,7 @@ export default defineConfig([
 	reactHooks.configs.flat.recommended,
 	{
 		rules: {
+			'@typescript-eslint/no-unused-vars': 'error',
 			'react/self-closing-comp': [
 				'error',
 				{
